[library][lg]: refactor datapub to be a library - #64
